cookbook/appworld/requirements-lock.txt
# AppWorld Benchmark Requirements
# Python 3.11 (required by AppWorld)
# Generated from .venv-appworld

aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==3.7.1
appworld==0.1.3
asttokens==3.0.1
async-generator==1.10
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==39.0.2
decorator==5.2.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
execnet==2.1.2
executing==2.2.1
faker==24.14.1
fastapi==0.110.3
fastapi-login==1.9.3
fastuuid==0.14.0
filelock==3.20.2
freezegun==1.2.2
frozenlist==1.8.0
fsspec==2025.12.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==1.2.4
idna==3.11
importlib-metadata==8.7.1
inflection==0.5.1
iniconfig==2.3.0
ipython==9.9.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
libcst==1.8.6
litellm==1.80.13
loguru==0.7.3
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
munch==4.0.0
numpy==2.4.0
openai==2.14.0
orjson==3.11.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pendulum==3.1.0
pexpect==4.9.0
pluggy==1.6.0
polyfactory==2.22.5
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.23
pydantic==1.10.26
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.10.1
pytest==8.4.2
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.9
pytz==2025.2
pyyaml==6.0.3
ray==2.53.0
referencing==0.37.0
regex==2025.11.3
registrable==0.0.4
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==1.4.54
sqlalchemy-utils==0.41.2
sqlalchemy2-stubs==0.0.2a38
sqlmodel==0.0.10
stack-data==0.6.3
starlette==0.37.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.21.0
typer==0.12.5
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.27.1
uvloop==0.19.0
wcwidth==0.2.14
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
